mfp/requirements.txt
# sudo pip3 install -r requirements.txt
Pillow==5.3.0
Django==1.11.7
djangorestframework==3.7.1
social-auth-app-django==1.2.0
django-celery-email==2.0.1
django-googledrive-storage==1.3.5
django-taggit==0.22.2
django-taggit-serializer==0.1.1
django-google-maps==0.10.1

